+++ CHANGELOG.md
@@ -28,6 +28,7 @@
 ### Changed
 
 * nostr: bump `bitcoin` to `v0.32` ([Yuki Kishimoto])
+* nostr: deprecate `Event::from_value` ([Yuki Kishimoto])
 * webln: bump `webln` to `v0.3` ([Yuki Kishimoto])
 * sdk: bump `lnurl-pay` to `v0.6` ([Yuki Kishimoto])